a2a_agents/backend/scripts/deployment/quick_deploy.py
#!/usr/bin/env python3
"""
Quick deployment script for A2A services
Starts all services with minimal configuration
"""

import subprocess
import time
import requests
import sys
import os

def check_port(port):
    """Check if a port is available"""
    try:
        response = requests.get(f"http://localhost:{port}/health", timeout=2)
        return response.status_code == 200
    except:
        return False

def start_service(name, port, module_path):
    """Start a service using uvicorn"""
    print(f"Starting {name} on port {port}...")
    
    # Create a minimal FastAPI app for each service
    app_code = f'''
import sys
sys.path.append("{os.path.dirname(os.path.abspath(__file__))}")

from fastapi import FastAPI
from fastapi.middleware.cors import CORSMiddleware

app = FastAPI(title="{name}")

app.add_middleware(
    CORSMiddleware,
    allow_origins=["*"],
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)

@app.get("/health")
async def health_check():
    return {{
        "status": "healthy",
        "service": "{name}",
        "port": {port}
    }}

@app.get("/trust/public-key")
async def get_public_key():
    return {{
        "public_key": "mock_public_key_for_{name}",
        "agent_id": "{name.lower().replace(' ', '_')}"
    }}

@app.post("/api/register")
async def register(data: dict):
    return {{"status": "success", "product_id": "test_123", "message": "Mock registration"}}

@app.post("/api/standardize")
async def standardize(data: dict):
    return {{"status": "success", "standardized_id": "std_123"}}

@app.post("/api/prepare")
async def prepare(data: dict):
    return {{"status": "success", "enriched": True}}

@app.post("/api/process-vectors")
async def process_vectors(data: dict):
    return {{"status": "success", "embeddings_created": True}}

@app.post("/api/validate-calculations")
async def validate_calculations(data: dict):
    return {{"status": "success", "passed": True}}

@app.post("/api/qa-validate")  
async def qa_validate(data: dict):
    return {{"status": "success", "score": 95}}

@app.post("/api/data")
async def store_data(data: dict):
    return {{"status": "success", "id": "data_123"}}

@app.post("/api/search")
async def search(data: dict):
    return {{"status": "success", "results": []}}

@app.post("/api/discover")
async def discover(data: dict):
    return {{"status": "success", "agents": ["{name}"]}}

@app.get("/api/blockchain/config")
async def blockchain_config():
    return {{
        "business_data_cloud_address": "0x9fE46736679d2D9a65F0992F2272dE9f3c7fa6e0",
        "agent_registry_address": "0x5FbDB2315678afecb367f032d93F642f64180aa3",
        "message_router_address": "0xe7f1725E7734CE288F8367e1Bb143E90bb3F0512"
    }}

@app.post("/trust/verify")
async def verify_trust(data: dict):
    return {{"status": "verified", "trust_level": 95}}

if __name__ == "__main__":
    import uvicorn
    uvicorn.run(app, host="0.0.0.0", port={port})
'''
    
    # Write the app to a temporary file
    filename = f"temp_{name.lower().replace(' ', '_')}.py"
    with open(filename, "w") as f:
        f.write(app_code)
    
    # Start the service
    process = subprocess.Popen(
        [sys.executable, filename],
        stdout=subprocess.PIPE,
        stderr=subprocess.PIPE
    )
    
    # Wait for service to start
    time.sleep(2)
    
    if check_port(port):
        print(f"✅ {name} started successfully on port {port}")
        return process
    else:
        print(f"❌ Failed to start {name}")
        process.terminate()
        return None

def main():
    print("🚀 A2A Quick Deployment")
    print("=" * 50)
    
    services = [
        ("Data Manager", 8001),
        ("Catalog Manager", 8002),
        ("Agent 0 - Data Product", 8003),
        ("Agent 1 - Standardization", 8004),
        ("Agent 2 - AI Preparation", 8005),
        ("Agent 3 - Vector Processing", 8008),
        ("Agent 4 - Calc Validation", 8006),
        ("Agent 5 - QA Validation", 8007),
    ]
    
    processes = []
    
    # Start all services
    for name, port in services:
        if check_port(port):
            print(f"⚠️  {name} already running on port {port}")
        else:
            process = start_service(name, port, "")
            if process:
                processes.append(process)
    
    print("\n✅ Deployment complete!")
    print("\nServices running at:")
    for name, port in services:
        print(f"  {name}: http://localhost:{port}")
    
    print("\n📋 To verify deployment:")
    print("  python verify_deployment.py")
    
    print("\n🧪 To run tests:")
    print("  python test_integration_quick.py")
    
    print("\n🛑 Press Ctrl+C to stop all services")
    
    try:
        # Keep the script running
        while True:
            time.sleep(1)
    except KeyboardInterrupt:
        print("\n\nStopping all services...")
        for process in processes:
            process.terminate()
        
        # Clean up temp files
        for name, _ in services:
            filename = f"temp_{name.lower().replace(' ', '_')}.py"
            if os.path.exists(filename):
                os.remove(filename)
        
        print("✅ All services stopped")

if __name__ == "__main__":
    main()